refactor: update internal utils imports
