+++ redbot/core/events.py
@@ -18,8 +18,8 @@ from .. import __version__ as red_version, version_info as red_version_info, Ver
 from . import commands
 from .config import get_latest_confs
 from .data_manager import storage_type
+from .utils._internal_utils import fuzzy_command_search, format_fuzzy_results
 from .utils.chat_formatting import inline, bordered, format_perms_list, humanize_timedelta
-from .utils import fuzzy_command_search, format_fuzzy_results
 
 log = logging.getLogger("red")
 init()
